+++ library/Toplevelview/Model/Hostgroupsummary.php
@@ -61,17 +61,17 @@ class Hostgroupsummary extends UnionModel
         return $q;
     }
 
-    public function getTableName()
+    public function getTableName(): string
     {
         return 'hostgroup';
     }
 
-    public function getKeyName()
+    public function getKeyName(): mixed
     {
         return ['id' => 'hostgroup_id'];
     }
 
-    public function getColumns()
+    public function getColumns(): array
     {
         return [
             'name'                        => 'hostgroup_name',
@@ -143,12 +143,12 @@ class Hostgroupsummary extends UnionModel
         ];
     }
 
-    public function getSearchColumns()
+    public function getSearchColumns(): array
     {
         return ['name_ci', 'display_name'];
     }
 
-    public function getDefaultSort()
+    public function getDefaultSort(): string
     {
         return 'display_name';
     }