+++ library/Toplevelview/Tree/TLVHostGroupNode.php
@@ -16,11 +16,9 @@ use stdClass;
  */
 class TLVHostGroupNode extends TLVIcingaNode
 {
-    protected $type = 'hostgroup';
-
-    protected $key = 'hostgroup';
-
-    protected static $titleKey = 'hostgroup';
+    protected string $type = 'hostgroup';
+    protected ?string $key = 'hostgroup';
+    protected static string $titleKey = 'hostgroup';
 
     public function getTitle(): string
     {
@@ -33,7 +31,7 @@ class TLVHostGroupNode extends TLVIcingaNode
             $n = $obj->display_name;
         }
 
-        return sprintf('%s', $n);
+        return $n;
     }
 
     public static function fetch(TLVTree $root): void
@@ -45,7 +43,7 @@ class TLVHostGroupNode extends TLVIcingaNode
         }
 
         $hgFilter = Filter::any();
-        foreach (array_keys($root->registeredObjects['hostgroup']) as $name) {
+        foreach ($root->registeredObjects['hostgroup'] as $name => $_) {
             $hgFilter->add(Filter::equal('hostgroup_name', $name));
         }
 
@@ -61,9 +59,7 @@ class TLVHostGroupNode extends TLVIcingaNode
             $hg->display_name = $hostgroup->display_name;
             $hg->hosts_total = $hostgroup->hosts_total;
             $hg->hosts_up = $hostgroup->hosts_up;
-            $hg->hosts_total = $hostgroup->hosts_total;
             $hg->services_total = $hostgroup->services_total;
-            $hg->hosts_up = $hostgroup->hosts_up;
             $hg->services_ok = $hostgroup->services_ok;
             $hg->hosts_down_handled = $hostgroup->hosts_down_handled;
             $hg->hosts_down_unhandled = $hostgroup->hosts_down_unhandled;